픽셀데이터 추출 1바이트만 추출해서 4바이트로 변경

+++ Source/Static/FEditorManager.cpp
@@ -298,7 +298,7 @@ FVector4 FEditorManager::GetPixel(FVector MPos) const
     FDevice::Get().GetDeviceContext()->Map(stagingTexture, 0, D3D11_MAP_READ, 0, &mapped);
 
     // 5. 픽셀 데이터 추출 (1x1 텍스처이므로 offset = 0)
-    const BYTE* pixelData = static_cast<const BYTE*>(mapped.pData);
+    const uint32* pixelData = static_cast<const uint32*>(mapped.pData);
 
     if (pixelData)
     {
